refactor(PersonPod): separate person pod container into its own file
