+++ app/components/ui/atoms/PersonPod.js
@@ -10,47 +10,8 @@ import ButtonAsIconWrapper from './ButtonAsIconWrapper'
 import SmallParagraph from './SmallParagraph'
 import Paragraph from './Paragraph'
 import PersonInfo from './PersonInfo'
+import PersonPodContainer from './PersonPodContainer'
 import ModalDialog from '../molecules/ModalDialog'
-
-const AddIcon = props => (
-  <Icon
-    iconName="Plus"
-    overrideName="@pubsweet-pending.PeoplePicker.PersonPod.Add"
-    {...props}
-  />
-)
-
-const StyledAddIcon = styled(AddIcon)`
-  fill: ${th('colorTextSecondary')};
-`
-
-const TrashIcon = props => (
-  <Icon
-    iconName="Trash"
-    overrideName="@pubsweet-pending.PeoplePicker.PersonPod.Remove"
-    {...props}
-  />
-)
-
-const StyledRemoveIcon = styled(TrashIcon)`
-  fill: ${th('colorTextSecondary')};
-`
-
-const CheckCircleIcon = props => (
-  <Icon
-    iconName="CheckCircle"
-    overrideName="@pubsweet-pending.PeoplePicker.PersonPod.Selected"
-    {...props}
-  />
-)
-
-const StyledSelectedIcon = styled(CheckCircleIcon)`
-  circle {
-    fill: ${th('colorPrimary')};
-  }
-  height: ${th('space.4')}
-  width: ${th('space.4')}
-`
 
 const InfoIcon = props => (
   <Icon iconName="Info" overrideName="info" {...props} />
@@ -61,22 +22,6 @@ const StyledInfoIcon = styled(InfoIcon)`
   height: 18px;
   width: 18px;
   fill: ${th('colorPrimary')};
-`
-
-const StyledButton = styled.button`
-  background-color: inherit;
-  height: 100%;
-  width: ${th('space.5')};
-  border: none;
-
-  &:hover ${StyledAddIcon}, &:hover ${StyledRemoveIcon} {
-      fill: #666666;
-    }
-
-  &:hover ${StyledSelectedIcon} > circle {
-      fill: #1378bb;
-    }
-  }
 `
 
 const StyledParagraph = styled(Paragraph)`
@@ -99,64 +44,9 @@ const SmallAction = styled(Action)`
   margin: 3px 0;
 `
 
-const StyledPod = styled(Flex)`
-  background: ${th('colorSecondary')};
-  border: ${th('borderWidth')} ${th('borderStyle')} ${th('colorBorder')};
-  border-radius: ${th('borderRadius')};
-  color: ${th('colorText')};
-  font-family: ${th('fontInterface')};
-  font-size: ${th('fontSizeBase')};
-  min-width: 312px;
-  height: 120px;
-`
-
 const CollapsibleBox = styled(Box)`
   min-width: 0;
 `
-
-const PersonPodContainer = ({
-  isSelectButtonClickable,
-  togglePersonSelection,
-  selectButtonType,
-  children,
-  ...props
-}) => (
-  <StyledPod justifyContent="space-between">
-    {children}
-    <Flex flexDirection="column" justifyContent="center">
-      {selectButtonType === 'remove' && (
-        <StyledButton
-          data-test-id="person-pod-button"
-          disabled={!isSelectButtonClickable}
-          onClick={togglePersonSelection}
-          type="button"
-        >
-          <StyledRemoveIcon />
-        </StyledButton>
-      )}
-      {selectButtonType === 'selected' && (
-        <StyledButton
-          data-test-id="person-pod-button"
-          disabled={!isSelectButtonClickable}
-          onClick={togglePersonSelection}
-          type="button"
-        >
-          <StyledSelectedIcon />
-        </StyledButton>
-      )}
-      {selectButtonType === 'add' && (
-        <StyledButton
-          data-test-id="person-pod-button"
-          disabled={!isSelectButtonClickable}
-          onClick={togglePersonSelection}
-          type="button"
-        >
-          <StyledAddIcon />
-        </StyledButton>
-      )}
-    </Flex>
-  </StyledPod>
-)
 
 class PersonPod extends React.Component {
   constructor(props) {
@@ -283,12 +173,6 @@ const ChooserPod = ({
     </Flex>
   </PersonPodContainer>
 )
-
-PersonPodContainer.propTypes = {
-  isSelectButtonClickable: PropTypes.bool.isRequired,
-  togglePersonSelection: PropTypes.func.isRequired,
-  selectButtonType: PropTypes.oneOf(['add', 'remove', 'selected']).isRequired,
-}
 
 PersonPod.propTypes = {
   isSelectButtonClickable: PropTypes.bool,
